requirements.txt
asgiref==3.4.1
Django==4.0
django-phone-field==1.8.1
Pillow==8.4.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
